WIP: Basic UI for villager advancement
Done:
- Show all jobs in UI (and lay them out nicely)
- Fix errors complaining about config being used too soon
- Added mouse-wheel scrolling to UI

TODO:
- Indicate "dining" and "seeking work" in UI
- Remove dirt cheat to open advancement UI
- 3 tiers of gatherer advancement
- Experience and blocks of progress

+++ src/main/java/ca/bradj/questown/core/Config.java
@@ -55,7 +55,7 @@ public class Config {
     public static final ForgeConfigSpec.ConfigValue<Boolean> CRASH_ON_FAILED_WARP;
     public static final ForgeConfigSpec.ConfigValue<Integer> TIME_WARP_MAX_TICKS;
     public static final ForgeConfigSpec.ConfigValue<Integer> BASE_MAX_LOOP;
-    public static final ForgeConfigSpec.ConfigValue<Integer> MAX_TICKS_WITHOUT_SUPPLIES;
+    public static final ForgeConfigSpec.ConfigValue<Long> MAX_TICKS_WITHOUT_SUPPLIES;
     public static final ForgeConfigSpec.ConfigValue<Integer> BASE_FULLNESS;
     public static final ForgeConfigSpec.ConfigValue<Boolean> HUNGER_ENABLED;
     public static final ForgeConfigSpec.ConfigValue<Long> BLOCK_CLAIMS_TICK_LIMIT;
@@ -124,7 +124,7 @@ public class Config {
         BUILDER.push("Jobs");
         MAX_TICKS_WITHOUT_SUPPLIES = BUILDER.comment(
                 "If the town is missing the supplies that the villager needs to do their job, they will wait some time for those supplies to be generated/added. After these ticks, they will give up and go back to the job board"
-        ).defineInRange("MaxTicksWithoutSupplies", 100, 1, 24000);
+        ).defineInRange("MaxTicksWithoutSupplies", 100L, 1L, 24000L);
         FARM_ACTION_INTERVAL = BUILDER.comment(
                 "The number of ticks that farmers will wait between actions"
         ).define("FarmActionInterval", 100);
